+++ SendBirdSDK.framework/Headers/SBDInternalTypes.h
@@ -73,4 +73,13 @@ typedef NS_ENUM(NSInteger, SBDLogLevel) {
     SBDLogLevelInfo = 3,
 };
 
+typedef NS_OPTIONS(NSUInteger, SBDChannelQueryIncludeOption) {
+    SBDChannelQueryIncludeOptionNone                    = (1 << 0),
+    SBDChannelQueryIncludeOptionEmptyChannel            = (1 << 1),
+    SBDChannelQueryIncludeOptionMemberList              = (1 << 2),
+    SBDChannelQueryIncludeOptionFrozenChannel           = (1 << 3),
+    SBDChannelQueryIncludeOptionReadReceipt             = (1 << 4),
+    SBDChannelQueryIncludeOptionDeliveryReceipt         = (1 << 5)
+};
+
 #endif /* SBDInternalTypes_h */